Interaction system migrations and improvements

+++ Technolithic/Technolithic/Components/Creatures/DomesticationData.cs
@@ -10,7 +10,7 @@ namespace Technolithic
     public class DomesticationData
     {
         public string TamedForm { get; set; }
-        public int Chance { get; set; }
+        public float Chance { get; set; }
 
         [JsonIgnore]
         public AnimalTemplate TamedFormAnimalTemplate { get; set; }
